cleaning up schemas and adding reply queries
